feat(Update profile information and view profile): Update profile, view profile
This is the two components for updating and viewing the profile. The update function is a put

request because we're assuming that admins will create users and then users will update their own

information, so there will be no need for a create new profile function.

Worked in shared redux files